fix: styles in the home page and video detail page
